requirements.txt
attr==0.3.2
beautifulsoup4==4.12.2
ConfigParser==6.0.0
cryptography==41.0.4
Cython==3.0.2
dl==0.1.0
docutils==0.20.1
feedparser==6.0.10
google_api_python_client==2.101.0
google_auth_oauthlib==1.1.0
HTMLParser==0.0.2
httplib2==0.22.0
ipython==8.16.1
ipywidgets==8.1.1
Jinja2==3.1.2
keyring==24.2.0
lockfile==0.12.2
moviepy==1.0.3
numpy==1.26.0
Pillow==10.0.1
playwright==1.38.0
protobuf==4.24.3
pydub==0.25.1
pyOpenSSL==23.2.0
python-dotenv==1.0.0
railroad==0.5.0
Sphinx==7.2.6
tornado==6.3.3
trove_classifiers==2023.9.19
truststore==0.8.0
urllib3_secure_extra==0.1.0
yt_dlp==2023.9.24
discord
pywa
tqdm
